Delete duplicate string of elements in linked list
Utilize the list_for_each_safe() function to compare the current node's value with the next node's value. If they are the same,
delete the current node. Upon reaching a different value, delete the previous node.
If the number of nodes is even, potentially delete the last node at the end to ensure adjacent duplicate nodes are removed.

Change-Id: I8c85f6d9ecf197ee0b41319cbf63690d07739edd