Implement functions of q_ascend and q_descend
First, set the node pointed to by head->prev as the maximum/minimum reference point. Then, traverse the list and compare each node.
If a node is smaller/larger than the reference point, delete that node. If a node is larger/smaller than the reference point, update the reference point to this node.

Change-Id: If6162552ad7bd2d8c24d57e02316dda462985d6b